Updated documentation for api batch and statics generation files
