+++ src/server/generate/batchAll.ts
@@ -1,10 +1,14 @@
-import * as batch from './batchMovies';
+import * as batch from './batchTradional';
 import * as batchAnime from './batchAnime';
 import * as batchGenre from './batchGenres';
 
 const sleep = (ms: number) => (
   new Promise((resolve) => setTimeout(resolve, ms)));
 
+/**
+ * Calls each of the external APIs to retrieve updated information about movies, tv shows,
+ * and anime.
+ */
 const allBatch = async () => {
   await batch.retrieveBatch('tv');
   await batch.retrieveBatch('movie');
