+++ src/server/generate/batchGenres.ts
@@ -1,6 +1,9 @@
 import * as fs from 'fs';
 import * as movieapi from '../services/movieService';
 
+/**
+ * Calls the external API for tv shows and movies and generates the genre list for these media types.
+ */
 export const retrieveBatch = async () => {
   const tvBatch = await movieapi.getGenres('tv');
   fs.writeFile('../api-json/tv_genres.json', JSON.stringify(tvBatch), (err : any) => {
